some modif

+++ java/tp2/test2.java
@@ -3,7 +3,7 @@ public class test2 {
     public static void main(String [] args){
         float a= 3.0;
         double b=4;
-        float c;
+        double c;
         c= Math.sqrt(a*a+b*b);
         System.out.println("c= "+ c);
